Change interface to not use stdin/stderr
In order to make it so that the config container can interract with the user
via stdin/stdout/stderr the interface is changing - the entrypoint will
handle user interaction and run a unix domain socket server for incoming
requests (requests and responses still in simple JSON format). cdflow2
will then communicate with this by docker exec'ing the config binary, which
will forward the request/response from/to the exec's stdin/stdout to the
socket service.